requirements.txt
fastapi
uvicorn
python-multipart
Pillow
boto3
